Create trace runner
