+++ src/analysis/mrc/generate_multi_mrc.c
@@ -10,15 +10,11 @@
 #include <string.h>
 
 #include "arrays/array_size.h"
-#include "evicting_map/evicting_map.h"
+#include "arrays/is_last.h"
 #include "file/file.h"
 #include "glib.h"
-#include "histogram/histogram.h"
 #include "logger/logger.h"
 #include "miss_rate_curve/miss_rate_curve.h"
-#include "olken/olken.h"
-#include "shards/fixed_rate_shards.h"
-#include "shards/fixed_size_shards.h"
 #include "timer/timer.h"
 #include "trace/generator.h"
 #include "trace/reader.h"
@@ -26,10 +22,7 @@
 
 #include "helper.h"
 #include "runner_arguments.h"
-
-/// @note   The keyword 'inline' prevents a compiler warning as per:
-///         https://stackoverflow.com/questions/32432596/warning-always-inline-function-might-not-be-inlinable-wattributes
-#define forceinline __attribute__((always_inline)) inline
+#include "trace_runner.h"
 
 struct CommandLineArguments {
     char *executable;
@@ -58,21 +51,6 @@ struct CommandLineArguments {
 
     bool cleanup;
 };
-
-/// @brief  Print algorithms by name in format: "{Olken,Fixed-Rate-SHARDS,...}".
-static void
-print_available_algorithms(FILE *stream)
-{
-    fprintf(stream, "{");
-    // NOTE We want to skip the "INVALID" algorithm name (i.e. 0).
-    for (size_t i = 1; i < ARRAY_SIZE(algorithm_names); ++i) {
-        fprintf(stream, "%s", algorithm_names[i]);
-        if (i != ARRAY_SIZE(algorithm_names) - 1) {
-            fprintf(stream, ",");
-        }
-    }
-    fprintf(stream, "}");
-}
 
 static struct CommandLineArguments
 parse_command_line_arguments(int argc, char *argv[])
@@ -204,22 +182,27 @@ free_command_line_arguments(struct CommandLineArguments *const args)
 static void
 print_command_line_arguments(struct CommandLineArguments const *args)
 {
-    LOGGER_INFO("CommandLineArguments(executable='%s', ...)", args->executable);
-    // fprintf(stderr,
-    //         "CommandLineArguments(executable='%s', input_path='%s', "
-    //         "algorithm='%s', output_path='%s', shards_ratio='%g', "
-    //         "artifical_trace_length='%zu', oracle_path='%s', "
-    //         "hist_num_bins=%zu, hist_bin_size=%zu, cleanup=%s)\n",
-    //         args->executable,
-    //         args->input_path,
-    //         algorithm_names[args->algorithm],
-    //         args->output_path,
-    //         args->shards_sampling_ratio,
-    //         args->artificial_trace_length,
-    //         args->oracle_path ? args->oracle_path : "(null)",
-    //         args->hist_num_bins,
-    //         args->hist_bin_size,
-    //         bool_to_string(args->cleanup));
+    fprintf(LOGGER_STREAM,
+            "CommandLineArguments(executable='%s', input='%s', format='%s', "
+            "length=%zu, oracle='%s', run=",
+            args->executable,
+            args->input_path,
+            algorithm_names[args->trace_format],
+            args->artificial_trace_length,
+            maybe_string(args->oracle));
+    if (args->run != NULL) {
+        fprintf(LOGGER_STREAM, "[");
+        for (size_t i = 0; args->run[i] != NULL; ++i) {
+            fprintf(LOGGER_STREAM, "'%s'", args->run[i]);
+            if (args->run[i + 1] != NULL) {
+                fprintf(LOGGER_STREAM, ",");
+            }
+        }
+        fprintf(LOGGER_STREAM, "]");
+    } else {
+        fprintf(LOGGER_STREAM, "null");
+    }
+    fprintf(LOGGER_STREAM, ")\n");
 }
 
 static void
@@ -227,178 +210,11 @@ print_trace_summary(struct CommandLineArguments const *args,
                     struct Trace const *const trace)
 {
     fprintf(LOGGER_STREAM,
-            "Trace(source='%s', format='%s', length=%zu)\n",
+            "Trace(source='%s', format='%s', trace=%p, length=%zu)\n",
             args->input_path,
             TRACE_FORMAT_STRINGS[args->trace_format],
+            (void *)trace->trace,
             trace->length);
-}
-
-/// @note   I forcibly inline this with the hope that the compiler will
-///         be able to realize that the function pointers are constants.
-///         I noticed an improvement from 8.2s to 7.6s on the Twitter
-///         trace, cluster15.bin. It did not fix the stackoverflow.
-static forceinline bool
-trace_runner(void *const runner_data,
-             struct RunnerArguments const *const args,
-             struct Trace const *const trace,
-             bool (*access_func)(void *const, uint64_t const),
-             bool (*postprocess_func)(void *const),
-             bool (*hist_func)(void *const, struct Histogram const **const),
-             void (*destroy_func)(void *const))
-{
-    struct MissRateCurve mrc = {0};
-    struct Histogram const *hist = NULL;
-
-    double t0 = get_wall_time_sec();
-    for (size_t i = 0; i < trace->length; ++i) {
-        // NOTE I really, really, really hope that the compiler is smart
-        //      enough to inline this function!!!
-        access_func(runner_data, trace->trace[i].key);
-        if (i % 1000000 == 0) {
-            LOGGER_TRACE("Finished %zu / %zu", i, trace->length);
-        }
-    }
-    double t1 = get_wall_time_sec();
-    postprocess_func(runner_data);
-    double t2 = get_wall_time_sec();
-    // NOTE We do NOT own the histogram data through the 'hist' object.
-    //      The 'runner_data' object maintains ownership of the data.
-    if (!hist_func(runner_data, &hist)) {
-        LOGGER_ERROR("histogram getter failed");
-        goto cleanup;
-    }
-    if (!MissRateCurve__init_from_histogram(&mrc, hist)) {
-        LOGGER_ERROR("MRC initialization failed");
-        goto cleanup;
-    }
-    double t3 = get_wall_time_sec();
-    LOGGER_INFO("%s -- Histogram Time: %f | Post-Process Time: %f | MRC Time: "
-                "%f | Total Time: %f",
-                algorithm_names[args->algorithm],
-                (double)(t1 - t0),
-                (double)(t2 - t1),
-                (double)(t3 - t2),
-                (double)(t3 - t0));
-    if (args->hist_path != NULL) {
-        if (!Histogram__save(hist, args->hist_path)) {
-            LOGGER_WARN("failed to save histogram in '%s'", args->hist_path);
-        }
-    }
-    if (args->mrc_path != NULL) {
-        if (!MissRateCurve__save(&mrc, args->mrc_path)) {
-            LOGGER_WARN("failed to save MRC in '%s'", args->mrc_path);
-        }
-    }
-
-    destroy_func(runner_data);
-    MissRateCurve__destroy(&mrc);
-    return true;
-cleanup:
-    destroy_func(runner_data);
-    MissRateCurve__destroy(&mrc);
-    return false;
-}
-
-static bool
-run_olken(struct RunnerArguments const *const args,
-          struct Trace const *const trace)
-{
-    struct Olken me = {0};
-    if (!Olken__init_full(&me,
-                          args->num_bins,
-                          args->bin_size,
-                          args->out_of_bounds_mode)) {
-        LOGGER_ERROR("initialization failed!");
-        return false;
-    }
-
-    return trace_runner(
-        &me,
-        args,
-        trace,
-        (bool (*)(void *const, uint64_t const))Olken__access_item,
-        (bool (*)(void *const))Olken__post_process,
-        (bool (*)(void *const,
-                  struct Histogram const **const))Olken__get_histogram,
-        (void (*)(void *const))Olken__destroy);
-}
-
-static bool
-run_fixed_rate_shards(struct RunnerArguments const *const args,
-                      struct Trace const *const trace)
-{
-    struct FixedRateShards me = {0};
-    if (!FixedRateShards__init_full(&me,
-                                    args->sampling_rate,
-                                    args->num_bins,
-                                    args->bin_size,
-                                    args->out_of_bounds_mode,
-                                    args->shards_adj)) {
-        LOGGER_ERROR("initialization failed!");
-        return false;
-    }
-
-    return trace_runner(
-        &me,
-        args,
-        trace,
-        (bool (*)(void *const, uint64_t const))FixedRateShards__access_item,
-        (bool (*)(void *const))FixedRateShards__post_process,
-        (bool (*)(void *const, struct Histogram const **const))
-            FixedRateShards__get_histogram,
-        (void (*)(void *const))FixedRateShards__destroy);
-}
-
-static bool
-run_fixed_size_shards(struct RunnerArguments const *const args,
-                      struct Trace const *const trace)
-{
-    struct FixedSizeShards me = {0};
-    if (!FixedSizeShards__init_full(&me,
-                                    args->sampling_rate,
-                                    args->max_size,
-                                    args->num_bins,
-                                    args->bin_size,
-                                    args->out_of_bounds_mode)) {
-        LOGGER_ERROR("initialization failed!");
-        return false;
-    }
-
-    return trace_runner(
-        &me,
-        args,
-        trace,
-        (bool (*)(void *const, uint64_t const))FixedSizeShards__access_item,
-        (bool (*)(void *const))FixedSizeShards__post_process,
-        (bool (*)(void *const, struct Histogram const **const))
-            FixedSizeShards__get_histogram,
-        (void (*)(void *const))FixedSizeShards__destroy);
-}
-
-static bool
-run_evicting_map(struct RunnerArguments const *const args,
-                 struct Trace const *const trace)
-{
-    struct EvictingMap me = {0};
-    if (!EvictingMap__init_full(&me,
-                                args->sampling_rate,
-                                args->max_size,
-                                args->num_bins,
-                                args->bin_size,
-                                args->out_of_bounds_mode)) {
-        LOGGER_ERROR("initialization failed!");
-        return false;
-    }
-
-    return trace_runner(
-        &me,
-        args,
-        trace,
-        (bool (*)(void *const, uint64_t const))EvictingMap__access_item,
-        (bool (*)(void *const))EvictingMap__post_process,
-        (bool (*)(void *const,
-                  struct Histogram const **const))EvictingMap__get_histogram,
-        (void (*)(void *const))EvictingMap__destroy);
 }
 
 /// @note   I introduce this function so that I can do perform some logic but
@@ -586,55 +402,6 @@ free_work_array(struct RunnerArgumentsArray *me)
     }
     free(me->data);
     *me = (struct RunnerArgumentsArray){0};
-}
-
-static bool
-run_runner(struct RunnerArguments const *const args,
-           struct Trace const *const trace)
-{
-    if (!args->ok) {
-        // NOTE I have a bunch of checks in place so this shouldn't
-        //      ever trigger unless someone calls this function from
-        //      another way.
-        LOGGER_WARN("skipping because it's not ok");
-        return false;
-    }
-    RunnerArguments__println(args, LOGGER_STREAM);
-    switch (args->algorithm) {
-    case MRC_ALGORITHM_OLKEN:
-        if (!run_olken(args, trace)) {
-            LOGGER_WARN("Olken failed");
-        }
-        return true;
-    case MRC_ALGORITHM_FIXED_RATE_SHARDS:
-        if (!run_fixed_rate_shards(args, trace)) {
-            LOGGER_WARN("Fixed-Rate SHARDS failed");
-        }
-        return true;
-    case MRC_ALGORITHM_FIXED_SIZE_SHARDS:
-        if (!run_fixed_size_shards(args, trace)) {
-            LOGGER_WARN("Fixed-Size SHARDS failed");
-        }
-        return true;
-    case MRC_ALGORITHM_EVICTING_MAP:
-        if (!run_evicting_map(args, trace)) {
-            LOGGER_WARN("Evicting Map failed");
-        }
-        return true;
-    case MRC_ALGORITHM_QUICKMRC:
-    case MRC_ALGORITHM_GOEL_QUICKMRC:
-    case MRC_ALGORITHM_AVERAGE_EVICTION_TIME:
-    case MRC_ALGORITHM_THEIR_AVERAGE_EVICTION_TIME:
-        LOGGER_WARN("not implemented algorithm %s",
-                    algorithm_names[args->algorithm]);
-        return false;
-    default:
-        LOGGER_WARN("invalid algorithm %s", algorithm_names[args->algorithm]);
-        fprintf(LOGGER_STREAM, "algorithms include: ");
-        print_available_algorithms(LOGGER_STREAM);
-        fprintf(LOGGER_STREAM, "\n");
-        return false;
-    }
 }
 
 static bool
