+++ src/analysis/mrc/trace_runner.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <stdbool.h>
+
+#include "runner_arguments.h"
+#include "trace/trace.h"
+
+bool
+run_runner(struct RunnerArguments const *const args,
+           struct Trace const *const trace);
